+++ frontend/src/pages/index.js
@@ -7,12 +7,12 @@ import { ToastContainer, toast, Zoom } from "react-toastify";
 import axios from "axios";
 import ReCAPTCHA from "react-google-recaptcha";
 import styles from "@/styles/Home.module.css";
-import Input from "@/components/Input/Input";
-import Button from "@/components/Button/Button";
-import Logo from "@/components/Logo/Logo";
-import SnsButtons from "@/components/SNS-buttons/sns-buttons";
-import TitleWeb from "@/assets/title/title-web";
-import TitleMobile from "@/assets/title/title-mobile";
+import Input from "@/components/ui/Input/Input";
+import Button from "@/components/ui/Button/Button";
+import Logo from "@/components/ui/Logo/Logo";
+import SnsButtons from "@/components/ui/SnsButtons/SnsButtons";
+import TitleWeb from "@/assets/title/TitleWeb";
+import TitleMobile from "@/assets/title/TitleMobile";
 
 const lexend = Lexend({
   subsets: ["latin"],
